Refactor SimulationPlotDialog to generic PlotDialog
